compress.c: implemented jumbo frames
The internals of LZ4_stream_t are no longer accessed, since the buffer
is no longer reallocated.  To compressed a jumbo frame, a separate
short-lived buffer is allocated.

The compressed size increases by 0.13% on average, among
/var/lib/apt/lists/*list.* (0.14% among /*pkglist.*, but 0.6% among
/ALT/Sisyphus/*/base/pkglist.classic+bloat).

This also fixes `total uncompressed size` field in the frame header.
It should include and now includes the implied leading magic bytes.

+++ zpkglist.h
@@ -35,7 +35,7 @@ extern "C" {
 // a string which describes the error.  Both strings normally come
 // from the read-only data section.
 int zpkglistCompress(int in, int out, const char *err[2],
-		     void (*hash)(void *buf, unsigned size, void *arg), void *arg)
+		     void (*hash)(void *buf, size_t size, void *arg), void *arg)
 		     __attribute__((nonnull(3)));
 
 // For decompression, a more general "Reader" API is provided.
